Stop overloading use of airflow_python_version
This variable was being used to both indicate the version of python
native libs to be installed, as well as telling the pip module which
python version to use, causing problems when we wanted to use the
current pathed python version but still install native libs correctly.

Thus, we change philosophy and say it's not up to this role to muck with
your python version at runtime, but rather assume that you've properly
set it up for this role beforehand.  In other words, we just use the
current pathed python for executing pip installs and it's up to the user
to pass in the appropriate python environment via `airflow_virtualenv`.